% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_barchart.R
\name{vis_barchart}
\alias{vis_barchart}
\title{Barchart of sports injuries per tissue and pathology type}
\usage{
vis_barchart(data)
}
\arguments{
\item{data}{Sports injury data, with columns Tissue, Pathology.Type and Frequency}
}
\value{
Coloured barchart of sports injury data at tissue and pathology type level
}
\description{
Generate a bar chart to show the frequency of sports injuries per tissue and pathology type.
}
\examples{
vis_barchart(injuryTissuePathTable)
}
