% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_heatmap_female.R
\name{vis_heatmap_female}
\alias{vis_heatmap_female}
\title{Heatmap of the female human body}
\usage{
vis_heatmap_female(data, body_view, low_colour, high_colour, title)
}
\arguments{
\item{data}{Sports injury data, with columns Body.area, Body.region and Frequency}

\item{body_view}{Anatomical view of female body (either front, back or side)}

\item{low_colour}{Low colour for gradient colouring of heatmap}

\item{high_colour}{High colour for gradient colouring of heatmap}

\item{title}{Title of heatmap}

\item{body_region}{Boolean variable indicating whether to display the data by body region. Default is to display data by body area and is FALSE}
}
\value{
Frequency heatmap of injury data on a female human body
}
\description{
Generate a heatmap of sports injury data onto a female human body with different anatomical views.
}
\examples{
vis_heatmap_female(injuryDataTable, "front", "yellow", "red", "Heat Map of Front Body")
}
