% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_heatmap_male.R
\name{vis_heatmap_male}
\alias{vis_heatmap_male}
\title{Heatmap of the male human body}
\usage{
vis_heatmap_male(
  data,
  body_view,
  low_colour,
  high_colour,
  title,
  body_region = FALSE,
  show_labels = FALSE,
  include_unspecified = TRUE,
  colourblind_friendly = FALSE,
  colourOption = "E"
)
}
\arguments{
\item{data}{Sports injury data, with columns Body.area, Body.region and Frequency}

\item{body_view}{Anatomical view of male body (either front/anterior, back/posterior or side/lateral)}

\item{low_colour}{Low colour for gradient colouring of heatmap}

\item{high_colour}{High colour for gradient colouring of heatmap}

\item{title}{Title of heatmap}

\item{body_region}{Boolean variable indicating whether to display the data by body region. Default is to display data by body area and is FALSE}

\item{show_labels}{Boolean variable indicating whether to display labels and frequency count for each body area/region}

\item{include_unspecified}{Boolean variable indicating whether to include frequency data with unspecified body area/region}

\item{colourblind_friendly}{Boolean variable indicating whether to use a colourblind friendly colour palette for heatmap gradient}

\item{colourOption}{Option to choose which colourblind friendly palette from the package viridis, either "A", "B", "C", "E", "F", "G" or "H" (Default is "E")}
}
\value{
Frequency heatmap of injury data on a male human body
}
\description{
Generate a heatmap of sports injury data onto a male human body with different anatomical views.
}
\examples{
vis_heatmap_male(injuryDataTable, "front", "yellow", "red", "Heat Map of Front Body", body_region = FALSE, show_labels = TRUE, include_unspecified = TRUE)
vis_heatmap_male(injuryDataTable, "back", "blue", "green", "Heat Map of Back Body", body_region = TRUE)
vis_heatmap_male(injuryDataTable, "side", "yellow", "red", "Heat Map of Side Body", include_unspecified = FALSE)
}
