% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_heatmap_male.R
\name{vis_heatmap_male}
\alias{vis_heatmap_male}
\title{Heatmap of the male human body}
\usage{
vis_heatmap_male(data, body_view, low_colour, high_colour, title)
}
\arguments{
\item{data}{Sports injury data, with columns Body.area, Body.region and Frequency}

\item{body_view}{Anatomical view of male body (either front, back or side)}

\item{low_colour}{Low colour for gradient colouring of heatmap}

\item{high_colour}{High colour for gradient colouring of heatmap}

\item{title}{Title of heatmap}
}
\value{
Frequency heatmap of injury data on a male human body
}
\description{
Generate a heatmap of sports injury data onto a male human body with different anatomical views.
}
\examples{
vis_heatmap_male(injuryDataTable, "front", "yellow", "red", "Heat Map of Front Body")
}
